// src/AddressPage/AddressForm/types.ts
export interface AddressFormValues {
  name: string;
  number: string;
  email: string;
  address: string;
  subdistrict: string;
  district: string;
  province: string;
  zipCode: string;
  note: string;
}

export const AddressFormValuesKeys = {
  name: "name",
  number: "number",
  email: "email",
  address: "address",
  subdistrict: "subdistrict",
  district: "district",
  province: "province",
  zipCode: "zipCode",
  note: "note",
};